Fixed all troubles with resmon.

+++ spyserv-c-api/Core/Dtos/DiskResultDto.cs
@@ -2,6 +2,6 @@ public class DiskResultDto
 {
     public string Type { get; } = ResMonType.Disk.ToString();
     public string Device { get; set; } = string.Empty;
-    public float ReadMbps { get; set; }
-    public float WriteMbps { get; set; }    
+    public double ReadMbps { get; set; }
+    public double WriteMbps { get; set; }    
 }
