+++ spyserv-c-api/Core/Dtos/MemoryResultDto.cs
@@ -1,6 +1,6 @@
 public class MemoryResultDto
 {
     public string Type { get; } = ResMonType.Memory.ToString();
-    public float UsedPercent { get; set; }
-    public float TotalMemoryMb { get; set; }
+    public double UsedPercent { get; set; }
+    public double TotalMemoryMb { get; set; }
 }
